refactor(routing): unify route-match preamble across pages and app routers
- routing: extract a shared generic matchRouteWithTrie<R> in
  route-matching.ts and use it from pages-router.ts (matchRoute) and
  app-router.ts (matchAppRoute). Both now share the same query-strip,
  trailing-slash, normalize, split, and trie-cache lookup preamble.

Pure refactor; no behavior change. Follow-up to #1058.

Note: the planned migration of pages applyGsspHeaders to the shared
mergeMiddlewareResponseHeaders helper was deferred — applyGsspHeaders
reads Node-style OutgoingHttpHeaders (Record<string, string | number |
boolean | string[]>) and joins multi-value non-cookie arrays with ", "
before set(), neither of which the shared Headers-based helper does.
Collapsing without preserving those semantics risks behavioral drift,
so it'll ship as a separate PR.
